requirements.txt
# Requirements automatically generated by pigar.
# https://github.com/Damnever/pigar

# app.py: 1
# feedparser == 5.2.1

# notif.py: 1,2,4
# gi == 1.2

# app.py: 10
python_telegram_bot == 10.0.1

# app.py: 4
# requests == 2.18.4
